refactor(#282): data class Operation 을 제거하고 SerializedOperation 을 사용하도록 변경
